+++ tests/Unit/HyperTest.php
@@ -18,6 +18,13 @@ use Rexlabs\HyperHttp\Message\Response;
 
 class HyperTest extends TestCase
 {
+    protected function setUp()
+    {
+        Hyper::setDefaultConfig([]);
+        Hyper::clearInstances();
+        Hyper::setDefaultLogger(new NullLogger());
+    }
+
     public function test_url()
     {
         $hyper = Hyper::make();
@@ -46,7 +53,7 @@ class HyperTest extends TestCase
         $hyper = Hyper::make([
             'guzzle' => [
                 'base_uri' => 'http://example.com/v1',
-            ]
+            ],
         ]);
 
         // No trailing slash on base
@@ -69,7 +76,7 @@ class HyperTest extends TestCase
 
     public function test_base_uri_from_subclass()
     {
-        $myHyper = new class extends Hyper {
+        $myHyper = new class() extends Hyper {
             protected static function getBaseUri()
             {
                 return 'http://example.com/v1';
@@ -345,7 +352,6 @@ class HyperTest extends TestCase
         $this->assertEquals('hello', $response->data->message);
         $this->assertEquals(5678, $response->data->id);
     }
-
 
     public function test_with_headers()
     {